requirements.txt
py-cord
google-generativeai
python-dotenv
flask
openai